(nearly) unaltered vers. of orig Fauzi bin prep script
Was: script in /work/meta4_bins/data/bins/fauzi/bins
called assemble_files.  Not under version control b/c of
.gitignore.

I moved it out to /work/meta4_bins/data/bins/assemble_files.sh
so I could re-fresh the bins, after noticing that the
updates to the bins I made with
/work/meta4_bins/data/bins/fauzi/raw/genome_bins/fix_names.sh
will be propagated into the copy we are operating on.